+++ src/models/chats/chat.message-handler.js
@@ -22,16 +22,19 @@ class ChatMessageHandler {
         this.chat = chat;
         tracker.onKegTypeUpdated(chat.id, 'message', this.onMessageDigestUpdate);
         this.onMessageDigestUpdate();
-        this._reactionsToDispose.push(reaction(() => this.chat.active && clientApp.isInChatsView, (active) => {
-            if (active) {
-                this.onMessageDigestUpdate();
-                this.markAllAsSeen();
-                this.removeMaker();
-            } else {
-                this.cancelTimers();
+        this._reactionsToDispose.push(reaction(
+            () => this.chat.active && clientApp.isInChatsView && clientApp.isReadingNewestMessages,
+            (active) => {
+                if (active) {
+                    this.onMessageDigestUpdate();
+                    this.markAllAsSeen();
+                    this.removeMaker();
+                } else {
+                    this.cancelTimers();
+                }
             }
-        }));
-        this._reactionsToDispose.push(reaction(() => socket.authenticated, (authenticated) => {
+        ));
+        this._reactionsToDispose.push(reaction(() => tracker.updatedAfterReconnect, (authenticated) => {
             if (authenticated) {
                 this.onMessageDigestUpdate();
             } else {
@@ -39,8 +42,11 @@ class ChatMessageHandler {
             }
         }));
         this._reactionsToDispose.push(reaction(
-            () =>
-                socket.authenticated && clientApp.isFocused && clientApp.isInChatsView && this.chat.active,
+            () => tracker.updatedAfterReconnect
+                && this.chat.active
+                && clientApp.isFocused
+                && clientApp.isInChatsView
+                && clientApp.isReadingNewestMessages,
             (userIsReading) => {
                 if (userIsReading) {
                     this.markAllAsSeen();
@@ -135,7 +141,11 @@ class ChatMessageHandler {
     }
 
     markAllAsSeen() {
-        if (!clientApp.isFocused || !clientApp.isInChatsView || !this.chat.active) return;
+        if (!clientApp.isFocused
+            || !clientApp.isInChatsView
+            || !this.chat.active
+            || !clientApp.isReadingNewestMessages
+        ) return;
         this._markAsSeenTimer = setTimeout(() => {
             this._markAsSeenTimer = null;
             if (!clientApp.isFocused || !clientApp.isInChatsView || !this.chat.active) return;