+++ CHANGELOG.md
@@ -2,6 +2,48 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="4.17.0"></a>
+# [4.17.0](https://github.com/PeerioTechnologies/peerio-icebear/compare/v4.16.6...v4.17.0) (2018-05-03)
+
+
+### Bug Fixes
+
+* **copy:** removed leading space ([292dff5](https://github.com/PeerioTechnologies/peerio-icebear/commit/292dff5))
+* changed title_grantAccess to button grantAccess ([3c87021](https://github.com/PeerioTechnologies/peerio-icebear/commit/3c87021))
+* messages marked as seen when not needed ([c65c399](https://github.com/PeerioTechnologies/peerio-icebear/commit/c65c399))
+
+
+### Features
+
+* added inviteNoWarning to contact store ([8571eb4](https://github.com/PeerioTechnologies/peerio-icebear/commit/8571eb4))
+
+
+
+<a name="4.16.6"></a>
+## [4.16.6](https://github.com/PeerioTechnologies/peerio-icebear/compare/v4.16.5...v4.16.6) (2018-05-02)
+
+
+### Bug Fixes
+
+* read receipts only send if isReadingNewestMessages ([#186](https://github.com/PeerioTechnologies/peerio-icebear/issues/186)) ([c9e41af](https://github.com/PeerioTechnologies/peerio-icebear/commit/c9e41af))
+
+
+
+<a name="4.16.5"></a>
+## [4.16.5](https://github.com/PeerioTechnologies/peerio-icebear/compare/v4.16.4...v4.16.5) (2018-05-02)
+
+
+
+<a name="4.16.4"></a>
+## [4.16.4](https://github.com/PeerioTechnologies/peerio-icebear/compare/v4.16.3...v4.16.4) (2018-05-01)
+
+
+
+<a name="4.16.3"></a>
+## [4.16.3](https://github.com/PeerioTechnologies/peerio-icebear/compare/v4.16.2...v4.16.3) (2018-05-01)
+
+
+
 <a name="4.16.2"></a>
 ## [4.16.2](https://github.com/PeerioTechnologies/peerio-icebear/compare/v4.16.1...v4.16.2) (2018-04-30)